210325 commit

+++ app/src/main/java/com/example/how_about_here/src/main/AccDetail/model/Rating.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 
 data class Rating(
     @SerializedName("avg(R.rating)")
-    val avgRRating: Int
+    val avgRRating: Double
 )
